Implement the ability to request XR compatibility on a WebGL context.
This change is one part of a series of changes to add multi-GPU support
for VR. This CL contains the changes in Blink to request XR
compatibility on a WebGL context and send the request to the browser
process. There are two ways to do this - either calling
makeXRCompatible() on an existing WebGL context or setting the
xrCompatible flag in the context attributes:
canvas.getContext('webgl', { xrCompatible: true })

makeXRCompatible() returns a promise and sends an asynchronous request
to the browser process. Specifying the xrCompatible attribute is
synchronous, as getContext() must return a WebGL context that is
already compatible.

This CL doesn't change any functionality - the browser process
immediately returns compatibility status regardless of whether we're
XR compatible. A subsequent change will implement restarting the GPU
process if needed to be compatible with the VR headset.

This code was originally reviewed as part of a larger CL here:
https://chromium-review.googlesource.com/c/chromium/src/+/2096778

getContext('webgl2', { xrCompatible: true}) will be added in crbug.com/1087356

Bug: 876140
Change-Id: Iab6bf321d062023e7996dd5ffa84612f69ade95c