Fix when atomic inline with self-painting layer is truncated
When |NGLineTruncator| sets |IsHiddenForPaint|, it does so by
cloning the original fragment. When hiding atomic inlines,
the fragment in the inline formatting context is cloned and
|IsHiddenForPaint| set, but the fragment cached in |LayoutBox|
is kept unchanged, without |IsHiddenForPaint|.

When such fragment has self-painting layer, the one in
|LayoutBox| is painted. To hide such fragments, we moved them
to outside of the clipping area, but this had visible side
effects to |scrollWidth|.

This patch stops moving them, but instead, check the fragment
in the inline formatting context when painting and hit-
testing.

Bug: 1005645
Change-Id: I8d1d55afd2a2f1b6c1b6672d5090476872cb4dee